chore: move formOptions to core, remove createFormFactory from Solid
