+++ packages/angular-form/src/index.ts
@@ -1,10 +1,12 @@
 export type {
-  FieldApi,
   FormOptions,
   FieldState,
   FieldValidateFn,
   FieldValidateAsyncFn,
 } from '@tanstack/form-core'
+
+export { FormApi, FieldApi, formOptions } from '@tanstack/form-core'
+
 export { injectForm } from './inject-form'
 export { TanStackField } from './tanstack-field.directive'
 export { injectStore } from './inject-store'
