+++ packages/form-core/src/index.ts
@@ -4,3 +4,4 @@ export * from './utils'
 export * from './util-types'
 export * from './types'
 export * from './mergeForm'
+export * from './formOptions'
