+++ packages/form-core/src/mergeForm.ts
@@ -1,5 +1,6 @@
 import type { FormApi } from './FormApi'
 import type { Validator } from './types'
+import type { NoInfer } from './util-types'
 
 export function mutateMergeDeep(target: object, source: object): object {
   const targetKeys = Object.keys(target)
