+++ packages/form-core/src/tests/formOptions.test.ts
@@ -1,0 +1,25 @@
+import { describe, expect, it } from 'vitest'
+import { FormApi, formOptions } from '../index'
+
+describe('formOptions', () => {
+  it('should allow default values to be set', async () => {
+    type Person = {
+      firstName: string
+      lastName: string
+    }
+
+    const formOpts = formOptions<Person>({
+      defaultValues: {
+        firstName: 'FirstName',
+        lastName: 'LastName',
+      },
+    })
+
+    const form = new FormApi({
+      ...formOpts,
+    })
+
+    expect(form.state.values['firstName']).toBe('FirstName')
+    expect(form.state.values['lastName']).toBe('LastName')
+  })
+})
