--- packages/react-form/src/formOptions.ts
@@ -1,10 +1,0 @@
-import type { FormOptions, Validator } from '@tanstack/form-core'
-
-export function formOptions<
-  TFormData,
-  TFormValidator extends Validator<TFormData, unknown> | undefined = undefined,
->(
-  defaultOpts?: FormOptions<TFormData, TFormValidator>,
-) {
-  return defaultOpts;
-}
