+++ packages/solid-form/src/index.ts
@@ -16,12 +16,14 @@ export type {
   ValidationMeta,
 } from '@tanstack/form-core'
 
-export { FormApi, FieldApi, functionalUpdate } from '@tanstack/form-core'
+export {
+  FormApi,
+  FieldApi,
+  functionalUpdate,
+  formOptions,
+} from '@tanstack/form-core'
 
 export { createForm } from './createForm'
 
 export type { CreateField, FieldComponent } from './createField'
 export { createField, Field } from './createField'
-
-export type { FormFactory } from './createFormFactory'
-export { createFormFactory } from './createFormFactory'
